==> src/expert_annotations/05_expert_annotations_data_for_templates.Rmd <==
---
title: "gs3_data_for_expert_annotation_version_1"
author: "Anna Steinberg"
date: "`r Sys.Date()`"
output: distill::distill_article
editor_options: 
  markdown: 
    wrap: 72
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Goal of the script: For data requiring expert demand, create a dataset containing values and units given by respective annotators

### Read data and packages

```{r}
library(tidyverse)
```

### Read in data

```{r}
expert_demand <- read.csv("../../data/processed/expert_annotations/expert_demand.csv")
cleaned_data <- read.csv("../../data/processed/non_expert_annotations/non_expert_annotations_cleaned.csv")
```

### Merge corrected value and corrected unit columns and additional columns annotated by annotators to LLM outputs

Pivot cleaned data to get - two columns with corrected value - two
columns with corrected unit - two columns with comment - two columns
with document comment - two columns with document expert needed for each
annotator of respective report

```{r}
source("../pivot.R")
```

```{r}
# Variable list can be extended any time
pivot_vars <- c("corrected_value", "corrected_unit", "page_correct", "corrected_page", "comment", "document_comment")
pivot_combined <- pivot_w_split(cleaned_data, pivot_vars)
```

Merge columns corrections to expert_demand

```{r}
expert_demand_merged <- expert_demand %>% 
  left_join(pivot_combined, by = c("report_name", "ID", "page_used", "llm_year", "llm_scope", "llm_value", "llm_unit", "annotator_ID_1", "annotator_ID_2"))
```

Create new column "document_expert_requested_by_at_least_one_annotator"
to indicate if at least one annotator requested an expert

```{r}
expert_demand_merged <- expert_demand_merged %>%
  mutate(document_expert_requested_by_at_least_one_annotator = ifelse(rowSums(select(., contains("document_expert_needed")) == "Yes", na.rm = TRUE) > 0, "Yes", "No"))
```

Create new column
"row_expert_requested_by_at_least_one_annotator_or_disagreement" to
indicate if at least one annotator requested an expert on a specific row
or there is a disagreement in that row (for color-coding in template)

*Comment: The column "expert_needed" should be used for this in the future instead of
following code because "expert_needed" contains also document-level
flagging and would color-code all rows of a report in template in case
of document-level flagging*

```{r}
expert_demand_merged <- expert_demand_merged %>%
  mutate(row_expert_requested_by_at_least_one_annotator_or_disagreement = ifelse(rowSums(select(., contains("record_expert_needed")) == "Yes", na.rm = TRUE) > 0 | agreement_value_and_unit == F, "Yes", "No"))
```

Select relevant columns

```{r}
expert_demand_merged <- expert_demand_merged %>%
  select(report_name, starts_with("document_comment"), document_expert_requested_by_at_least_one_annotator, row_expert_requested_by_at_least_one_annotator_or_disagreement, ID, page_used, llm_year, llm_scope, llm_value, starts_with("value_correct"), starts_with("corrected_value"), llm_unit, starts_with("unit_correct"), starts_with("corrected_unit"), starts_with("corrected_page"), starts_with("comment"))
```

Save the dataset

```{r}
write.csv(expert_demand_merged, "../../data/processed/expert_annotations/data_for_expert_annotation_templates.csv", row.names = FALSE)
```
